refeactorized BE
